+++ examples/example.php
@@ -30,7 +30,6 @@ $bom->getComponentRepository()->addComponent(
     new \CycloneDX\Core\Models\Component(
         \CycloneDX\Core\Enums\Classification::LIBRARY,
         'myComponent',
-        '1.33.7'
     )
 );